+++ changelog.md
@@ -23,6 +23,8 @@
 - Added class `Chats/Channels/Attachments/Links`
 - Added class `Chats/Channels/Attachments/Videos`
 - Added class `Calls`
+- Added class `Checklists`
+- Added class `Checklists/Items`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`